feat: admin functionality

+++ routes/web.php
@@ -17,11 +17,14 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
+Route::controller(LoginController::class)->group(function () {
+    Route::get('logout', 'logout')->name('logout');
+});
+
 Route::middleware('guest')->group(function () {
     Route::controller(LoginController::class)->group(function () {
         Route::get('login', 'index')->name('login');
         Route::post('login', 'login');
-        Route::get('logout', 'logout')->name('logout');
     });
 
     Route::controller(RegisterController::class)->group(function () {
@@ -31,7 +34,7 @@ Route::middleware('guest')->group(function () {
 });
 
 Route::middleware('auth')->group(function () {
-    Route::middleware('not_role:admin,/admin')->group(function () {
+    Route::middleware('not_role:admin,admin.books.index')->group(function () {
         Route::prefix('category')->group(function () {
             Route::controller(User\CategoryController::class)->group(function () {
                 Route::get('/', 'index')->name('categories.index');
@@ -46,29 +49,31 @@ Route::middleware('auth')->group(function () {
         Route::controller(User\BookController::class)->group(function () {
             Route::get('/', 'index')->name('books.index');
             Route::post('/', 'store')->name('books.store');
-            Route::get('/{id}', 'show')->name('books.show');
-            Route::post('/{id}', 'update')->name('books.update');
-            Route::delete('/{id}', 'destroy')->name('books.destroy');
+            Route::get('/book/{id}', 'show')->name('books.show');
+            Route::post('/book/{id}', 'update')->name('books.update');
+            Route::delete('/book/{id}', 'destroy')->name('books.destroy');
         });
     });
 
     Route::prefix('admin')->group(function () {
-        Route::middleware('role:admin')->group(function () {
+        Route::middleware(['role:admin'])->group(function () {
             Route::prefix('category')->group(function () {
                 Route::controller(Admin\CategoryController::class)->group(function () {
                     Route::get('/', 'index')->name('admin.categories.index');
                     Route::post('/', 'store')->name('admin.categories.store');
                     Route::post('/{id}', 'update')->name('admin.categories.update');
                     Route::delete('/{id}', 'destroy')->name('admin.categories.destroy');
+
+                    Route::get('/datatable', 'datatable')->name('admin.categories.datatable');
                 });
             });
 
             Route::controller(Admin\BookController::class)->group(function () {
                 Route::get('/', 'index')->name('admin.books.index');
                 Route::post('/', 'store')->name('admin.books.store');
-                Route::get('/{id}', 'show')->name('admin.books.show');
-                Route::post('/{id}', 'update')->name('admin.books.update');
-                Route::delete('/{id}', 'destroy')->name('admin.books.destroy');
+                Route::get('/book/{id}', 'show')->name('admin.books.show');
+                Route::post('/book/{id}', 'update')->name('admin.books.update');
+                Route::delete('/book/{id}', 'destroy')->name('admin.books.destroy');
             });
         });
     });
